CALS-5887 Jobs support for solving connx pagination problem
